Add new file for QuestionsReducer
