+++ src/reducers/index.js
@@ -1,10 +1,12 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 import { foldersReducer, foldersLoadingReducer, foldersFailReducer } from './foldersReducer';
+import { questionsReducer } from './questionsReducer';
 
 export default combineReducers({
   router: routerReducer,
   folders: foldersReducer,
   foldersLoading: foldersLoadingReducer,
   foldersFail: foldersFailReducer,
+  questionsReducer,
 });
